// src/constants/types/main-content.const.ts
export enum ContentHeaderHandler {
    reset = "content-header/reset",
    search = "content-header/search",
    submit = "content-header/submit",
    close = "content-header/close",
    link = "content-header/link",
    update = "content-header/update"
}

export enum ContentFooterHandler {
    reset = "content-footer/reset",
    search = "content-footer/search",
    submit = "content-footer/submit",
    close = "content-footer/close",
    link = "content-footer/link",
    update = "content-footer/update"
}

export const enum theme {
    default = "default",
    popup = "popup",
}